chore(scripts): add utility getTypesSource
